chore(Navigation): rewrite reducer to slice [YTFRONT-3869]
