--- packages/ui/src/ui/constants/navigation/modals/table-sort-modal.ts
@@ -1,1 +1,0 @@
-export const TABLE_SORT_MODAL_PARTIAL = 'TABLE_SORT_MODAL_PARTIAL';
